+++ src/backend_cairo/mod.rs
@@ -350,9 +350,13 @@ fn render_group_impl(
     if let Some(ref id) = g.filter {
         if let Some(filter_node) = node.tree().defs_by_id(id) {
             if let usvg::NodeKind::Filter(ref filter) = *filter_node.borrow() {
-                let background = prepare_filter_background(node, filter, opt);
                 let ts = usvg::Transform::from_native(&curr_ts);
-                filter::apply(filter, bbox, &ts, opt, background.as_ref(), &mut *sub_surface);
+                let background = prepare_filter_background(node, filter, opt);
+                let fill_paint = prepare_filter_fill_paint(node, filter, bbox, ts, opt, &sub_surface);
+                let stroke_paint = prepare_filter_stroke_paint(node, filter, bbox, ts, opt, &sub_surface);
+                filter::apply(filter, bbox, &ts, opt,
+                              background.as_ref(), fill_paint.as_ref(), stroke_paint.as_ref(),
+                              &mut *sub_surface);
             }
         }
     }
@@ -416,6 +420,62 @@ fn prepare_filter_background(
     // Render from the `start_node` until the `parent`. The `parent` itself is excluded.
     let mut state = RenderState::RenderUntil(parent.clone());
     render_node_to_canvas_impl(&start_node, opt, view_box, img_size, &mut state, &cr);
+
+    Some(surface)
+}
+
+/// Renders an image used by `FillPaint`/`StrokePaint` filter input.
+///
+/// FillPaint/StrokePaint is mostly an undefined behavior and will produce different results
+/// in every application.
+/// And since there are no expected behaviour, we will simply fill the filter region.
+///
+/// https://github.com/w3c/fxtf-drafts/issues/323
+fn prepare_filter_fill_paint(
+    parent: &usvg::Node,
+    filter: &usvg::Filter,
+    bbox: Option<Rect>,
+    ts: usvg::Transform,
+    opt: &Options,
+    canvas: &cairo::ImageSurface,
+) -> Option<cairo::ImageSurface> {
+    let region = crate::filter::calc_region(filter, bbox, &ts, canvas).ok()?;
+    let surface = create_subsurface(region.size())?;
+    if let usvg::NodeKind::Group(ref g) = *parent.borrow() {
+        if let Some(paint) = g.filter_fill.clone() {
+            let cr = cairo::Context::new(&*surface);
+            let style_bbox = bbox.unwrap_or_else(|| Rect::new(0.0, 0.0, 1.0, 1.0).unwrap());
+            let fill = Some(usvg::Fill::from_paint(paint));
+            style::fill(&parent.tree(), &fill, opt, style_bbox, &cr);
+            cr.rectangle(0.0, 0.0, region.width() as f64, region.height() as f64);
+            cr.paint();
+        }
+    }
+
+    Some(surface)
+}
+
+/// The same as `prepare_filter_fill_paint`, but for `StrokePaint`.
+fn prepare_filter_stroke_paint(
+    parent: &usvg::Node,
+    filter: &usvg::Filter,
+    bbox: Option<Rect>,
+    ts: usvg::Transform,
+    opt: &Options,
+    canvas: &cairo::ImageSurface,
+) -> Option<cairo::ImageSurface> {
+    let region = crate::filter::calc_region(filter, bbox, &ts, canvas).ok()?;
+    let surface = create_subsurface(region.size())?;
+    if let usvg::NodeKind::Group(ref g) = *parent.borrow() {
+        if let Some(paint) = g.filter_stroke.clone() {
+            let cr = cairo::Context::new(&*surface);
+            let style_bbox = bbox.unwrap_or_else(|| Rect::new(0.0, 0.0, 1.0, 1.0).unwrap());
+            let fill = Some(usvg::Fill::from_paint(paint));
+            style::fill(&parent.tree(), &fill, opt, style_bbox, &cr);
+            cr.rectangle(0.0, 0.0, region.width() as f64, region.height() as f64);
+            cr.paint();
+        }
+    }
 
     Some(surface)
 }