+++ src/backend_skia/mod.rs
@@ -286,7 +286,11 @@ fn render_group_impl(
             if let usvg::NodeKind::Filter(ref filter) = *filter_node.borrow() {
                 let ts = usvg::Transform::from_native(&curr_ts);
                 let background = prepare_filter_background(node, filter, opt);
-                filter::apply(filter, bbox, &ts, opt, background.as_ref(), &mut sub_surface);
+                let fill_paint = prepare_filter_fill_paint(node, filter, bbox, ts, opt, &sub_surface);
+                let stroke_paint = prepare_filter_stroke_paint(node, filter, bbox, ts, opt, &sub_surface);
+                filter::apply(filter, bbox, &ts, opt,
+                              background.as_ref(), fill_paint.as_ref(), stroke_paint.as_ref(),
+                              &mut sub_surface);
             }
         }
     }
@@ -345,6 +349,58 @@ fn prepare_filter_background(
     render_node_to_canvas_impl(&start_node, opt, view_box, img_size, &mut state, &mut img);
 
     Some(img)
+}
+
+/// Renders an image used by `FillPaint`/`StrokePaint` filter input.
+///
+/// FillPaint/StrokePaint is mostly an undefined behavior and will produce different results
+/// in every application.
+/// And since there are no expected behaviour, we will simply fill the filter region.
+///
+/// https://github.com/w3c/fxtf-drafts/issues/323
+fn prepare_filter_fill_paint(
+    parent: &usvg::Node,
+    filter: &usvg::Filter,
+    bbox: Option<Rect>,
+    ts: usvg::Transform,
+    opt: &Options,
+    canvas: &skia::Surface,
+) -> Option<skia::Surface> {
+    let region = crate::filter::calc_region(filter, bbox, &ts, canvas).ok()?;
+    let mut surface = create_subimage(region.size())?;
+    if let usvg::NodeKind::Group(ref g) = *parent.borrow() {
+        if let Some(paint) = g.filter_fill.clone() {
+            let style_bbox = bbox.unwrap_or_else(|| Rect::new(0.0, 0.0, 1.0, 1.0).unwrap());
+            let fill = Some(usvg::Fill::from_paint(paint));
+            let fill = style::fill(&parent.tree(), &fill, opt, style_bbox, ts);
+            surface.draw_rect(0.0, 0.0, region.width() as f64, region.height() as f64, &fill);
+        }
+    }
+
+    Some(surface)
+}
+
+/// The same as `prepare_filter_fill_paint`, but for `StrokePaint`.
+fn prepare_filter_stroke_paint(
+    parent: &usvg::Node,
+    filter: &usvg::Filter,
+    bbox: Option<Rect>,
+    ts: usvg::Transform,
+    opt: &Options,
+    canvas: &skia::Surface,
+) -> Option<skia::Surface> {
+    let region = crate::filter::calc_region(filter, bbox, &ts, canvas).ok()?;
+    let mut surface = create_subimage(region.size())?;
+    if let usvg::NodeKind::Group(ref g) = *parent.borrow() {
+        if let Some(paint) = g.filter_stroke.clone() {
+            let style_bbox = bbox.unwrap_or_else(|| Rect::new(0.0, 0.0, 1.0, 1.0).unwrap());
+            let fill = Some(usvg::Fill::from_paint(paint));
+            let fill = style::fill(&parent.tree(), &fill, opt, style_bbox, ts);
+            surface.draw_rect(0.0, 0.0, region.width() as f64, region.height() as f64, &fill);
+        }
+    }
+
+    Some(surface)
 }
 
 fn create_layers(